Ensure onNewIntent() passes NEW intent extras
There is no way to mark an Intent as resolved / consumed.

If the user tries to send/view to the same app multiple times in a row without force
quitting the app in between, the same EXTRAs will be returned.

To work around this, update the Activity's intent for each new intent.